==> 1.scan.ipynb <==
pip install -r requirements.txt

from github import Github

# Authentication is defined via github.Auth
from github import Auth
import pandas as pd
import os 



department = 'BLC'
private_token = ''
host = "https://github.com"

auth = Auth.Token(private_token)
g = Github(auth=auth)
g.get_user().login

info_folder = os.path.join('info', department)
os.makedirs(info_folder, exist_ok=True)  # Tạo thư mục nếu chưa tồn tại

# 1. Lấy danh sách repositories mà người dùng sở hữu
user = g.get_user()
repos = user.get_repos()

list_repos = []
for repo in repos:
    list_repos.append({
        'id': repo.id,
        'name': repo.name,
        'full_name': repo.full_name,
        'owner': repo.owner.login,
        'created_at': repo.created_at,
        'default_branch': repo.default_branch,
    })

df_repos = pd.DataFrame(list_repos)

df_repos

file_repos = os.path.join(info_folder, 'repos.csv')
df_repos.to_csv(file_repos, index=False)

orgs = user.get_orgs()
list_orgs = []
for org in orgs:
    list_orgs.append({
        'id': org.id,
        'name': org.login,
        'full_name': org.name,
        'created_at': org.created_at,
    })

df_orgs = pd.DataFrame(list_orgs)
df_orgs

file_orgs = os.path.join(info_folder, 'orgs.csv')
df_orgs.to_csv(file_orgs, index=False)

list_team_members = []

for org in orgs:
    teams = org.get_teams()
    for team in teams:
        members = team.get_members()
        for member in members:
            list_team_members.append({
                'org_name': org.login,
                'team_id': team.id,
                'team_name': team.name,
                'member_id': member.id,
                'member_login': member.login,
                'member_name': member.name,
                'created_at': member.created_at,
            })

df_team_members = pd.DataFrame(list_team_members)
df_team_members

file_team_members = os.path.join(info_folder, 'team_members.csv')
df_team_members.to_csv(file_team_members, index=False)

list_repo_members = []

for repo in repos:
    try:
        collaborators = repo.get_collaborators()  # Lấy tất cả các cộng tác viên của repo
        for collaborator in collaborators:
            print(collaborator)
            list_repo_members.append({
                'repo_name': repo.name,
                'repo_full_name': repo.full_name,
                'member_id': collaborator.id,
                'member_login': collaborator.login,
                'member_name': collaborator.name,
                'permissions': collaborator.permissions,
                'created_at': collaborator.created_at,
            })
    except Exception as ex:
        print(f"Error fetching collaborators for {repo.full_name}: {ex}")

df_repo_members = pd.DataFrame(list_repo_members)
df_repo_members


file_repo_members = os.path.join(info_folder, 'repo_members.csv')
df_repo_members.to_csv(file_repo_members, index=False)
