==> 3.import.ipynb <==
pip install -r requirements.txt

import gitlab
import pandas as pd
import os
import time

department = "vnpay"
host = 'http://localhost:10001'
private_token = ''
info_folder = os.path.join('info', department)

gl = gitlab.Gitlab(url=host, private_token=private_token)
gl.auth()

username =gl.user.username

try:
    zgroup = gl.groups.create({'name': department, 'path': department})
except Exception as ex:
    print(ex)
group = gl.groups.get(department)
parent_id = group.id

cfg_group_file = f'info/{department}/organization-export.csv'
df_group = pd.read_csv(cfg_group_file, index_col=0)
df_group.head()

for index, group in df_group.iterrows():
    print(group['name'])
    group_name = group['name']
    try:
        output = gl.groups.create({
            'name': group_name,
            'path': group_name,
            'parent_id': parent_id,
        })
        print(f"Done create group {group_name}")
    except Exception as ex:
        print(f"Error while importing group {group_name}: {ex}")

cfg_project_file = f'info/{department}/repository-export.csv'
df_project = pd.read_csv(cfg_project_file, index_col=0)
df_project.head()

group = gl.groups.get(department)
projects = df_project.to_dict('records')

host_with_user = host.replace("//", f"//{username}:{private_token}@")

namespaces = gl.namespaces.list(get_all=True)
list_groups = []
for namespace in namespaces:
   list_groups.append({
         'id': namespace.id,
         # 'name': dgroup.name,
         # 'path': dgroup.path,
         # 'full_name': dgroup.full_name,
         'full_path': namespace.full_path,
         # 'parent_id': dgroup.parent_id,
         # 'created_at': dgroup.created_at,
         # 'organization_id': dgroup.organization_id,
    })
# print(list_groups)
list_groups

def get_fullpath(name, name_with_namespace):
    # print(name_with_namespace)
    parts = name_with_namespace.split('/')
    parts.pop()
    # print(parts)
    parts.insert(0, name)
    return '/'.join(str(x).strip() for x in parts) 
def get_parent(groups, project_path):
    for group in groups:
        if group['full_path'] == project_path: 
            return group['id']
    return None 

def found_project(project_name_with_namespace):
    try:
      return gl.projects.get(project_name_with_namespace)
    except:
      return None 

def create_new_project(project_name, parent_id):
    try:
        # Tạo repository mới
        new_project = gl.projects.create(
            name=project_name,
            namespace_id=parent_id,
            visibility='private',
        )
        return new_project
    except Exception as e:
        print(f"Error while creating project {project_name}: {e}")
        return None

import_status = []
for project in projects: 
    project_file = project['project_file']
    project_path = get_fullpath(department, project['name_with_namespace'])
    project_name = project['name']
    print(f"Start importing project: {project_name}, path: {project_path}")
    
    project_name_with_namespace = f"{project_path}/{project_name}"
    project = found_project(project_name_with_namespace)

    if project: 
        print(f"Project: {project_name_with_namespace} exists")
        import_status.append({
            'name': project_name,
            'name_with_namespace': project_name_with_namespace,
            'project_url': f"{host}/{project_name_with_namespace}",
            'ex': f"Project: {project_name_with_namespace} exists",
            'status': "exists"
        })
        continue 
    
    parent_id = get_parent(list_groups, project_path)
    if not parent_id: 
        parent_id = group.id  # Default in department group 
        print(f"Project: {project_name} not found parent id")
    
    try:
        # Tạo project mới trên GitLab hoặc hệ thống Git của bạn
        new_project = gl.projects.create({
            'name': project_name,
            'namespace_id': parent_id,
            'visibility': 'private'
        })
        print(f"Push {project_file} to {host}/{project_name_with_namespace}")
        
        # Lệnh push tất cả các nhánh và tags lên remote mới
        os.system(f"cd {project_file} && git push --all {host_with_user}/{project_name_with_namespace}")
        os.system(f"cd {project_file} && git push --tags {host_with_user}/{project_name_with_namespace}")
        import_status.append({
            'name': project_name,
            'name_with_namespace': project_name_with_namespace,
            'project_url': f"{host}/{project_name_with_namespace}",
            'ex': f'{None}',
            'status': "success"
        })
    except Exception as e:
        import_status.append({
            'name': project_name,
            'name_with_namespace': project_name_with_namespace,
            'project_url': f"{host}/{project_name_with_namespace}",
            'ex': f'{ex}',
            'status': "failed"
        })
        print(f"An unexpected error occurred while creating project {project_name}: {e}")
    
print("Done pushing all branches and tags to the new remote.")

import_status

df = pd.DataFrame(import_status)
df
